Add in padUsersCount method and API call
The PadMessageHandler objects now have a new API call associated
with them. I'm sure that's a funny place to put it, but the
pad2sessions object in that file seems to be the only place user
counts are stored!

Anyway, I hope this is helpful. I know it would be for me :)