Added the License to the code in order to runs under LibreJS standards
Please read how to free your javascript http://www.gnu.org/software/librejs/free-your-javascript.html

tested using Firefox 14.0.1 and LibreJS 4.8 add-on

I added Google Inc as copyright holder. There is an open issue on this:

https://github.com/Pita/etherpad-lite/issues/935

Is not clear to me if it should be 'Peter Martischka and contributors' or 'Google Inc.' or somebody else.