Move `title` attribute of buttons from `<span>` to `<a>`
The `.toolbar ul li a` that holds the `span.button` has a
padding of `4px 5px`. This means if the mouse is on the edge
of the button the `:hover` state is active but the relevant
tooltip does not show up. Moving the `title` attribute to
the `<a>` itself solves this problem.

Also removed some extra whitespace.