Add a subset of FontAwesome icons using Fontello
Most of the code here has been taken from @LaKing's #2554.

Only icons required for etherpad have been added. Here is a
list of the code points that are being used:

 0xe800 0xe801 0xe802 0xe803
 0xe804 0xe805 0xe806 0xe807
 0xe808 0xe809 0xe80a 0xe80b
 0xe80c 0xe80d 0xe80e 0xe80f
 0xe810 0xe811 0xe812 0xe813
 0xe814 0xe815 0xe816 0xe817

Their related icon names can be seen in:
/src/static/css/fontawesome-etherpad.css

Even though this patch *adds* all the required icons it only
uses them in the in toolbar. The switch to using only FontAwesome
icons and the removal of old PNG files will only happen after the
timeslider interface has been updated.

This is the first step for #2253