[chore] Allow debug mode on node versions >= 6.3
`node-inspector` is not supported on newer versions of Node. According
to the documentation of that tool, "Since version 6.3, Node.js provides
a built-in DevTools-based debugger which mostly deprecates Node
Inspector (...). The built-in debugger (...) provides certain advanced
features (...) that are too difficult to implement in Node Inspector."

Allow external connections to the inspector, so Etherpad instances
running on containers can also be debugged.